requirements.txt
Django
pillow
django-jazzmin
django-tinymce